+++ ginga/qtw/ColorBar.py
@@ -1,6 +1,6 @@
 #
 # ColorBar.py -- color bar widget
-# 
+#
 # Eric Jeschke (eric@naoj.org)
 #
 # Copyright (c) Eric R. Jeschke.  All rights reserved.
@@ -28,19 +28,19 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
         self.logger = logger
         self.pixmap = None
         self.link_rgbmap = link
-        
+
         if not rgbmap:
             rgbmap = RGBMap.RGBMapper(logger)
         self.set_rgbmap(rgbmap)
-        
+
         self._start_x = 0
         # for drawing range
         self.t_showrange = True
         self.t_font = 'Sans Serif'
         self.t_fontsize = 8
         self.t_spacing = 40
-        self.loval = 0
-        self.hival = 0
+        self.loval = 0.0
+        self.hival = 0.0
         self._interval = {}
         self._avg_pixels_per_range_num = 70
 
@@ -57,7 +57,7 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
 
     def get_rgbmap(self):
         return self.rgbmap
-        
+
     def set_rgbmap(self, rgbmap):
         self.rgbmap = rgbmap
         # TODO: figure out if we can get rid of this link option
@@ -74,12 +74,12 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
         self.redraw()
 
     def set_range(self, loval, hival, redraw=True):
-        self.loval = loval
-        self.hival = hival
+        self.loval = float(loval)
+        self.hival = float(hival)
         # Calculate reasonable spacing for range numbers
         if self.pixmap is not None:
             cr = self.setup_cr()
-            text = "%d" % (int(hival))
+            text = "%.4g" % (hival)
             rect = cr.boundingRect(0, 0, 1000, 1000, 0, text)
             x1, y1, x2, y2 = rect.getCoords()
             _wd = x2 - x1
@@ -90,7 +90,7 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
 
             if self.t_showrange and redraw:
                 self.redraw()
-        
+
     def get_size(self):
         rect = self.geometry()
         x1, y1, x2, y2 = rect.getCoords()
@@ -98,7 +98,7 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
         width = x2 - x1
         height = y2 - y1
         return width, height
-    
+
     def resizeEvent(self, event):
         width, height = self.get_size()
         #print "making pixmap of %dx%d" % (width, height)
@@ -115,7 +115,7 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
             nums, spacing, self._interval))
 
         self.redraw()
-       
+
     def paintEvent(self, event):
         """When an area of the window is exposed, we just copy out of the
         server-side, off-screen pixmap to that area.
@@ -127,7 +127,7 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
 
     def sizeHint(self):
         return QtCore.QSize(100, 16)
-    
+
     def repaint(self, rect):
         x1, y1, x2, y2 = rect.getCoords()
         width = x2 - x1
@@ -138,7 +138,7 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
         painter = QPainter(self)
         rect = QtCore.QRect(x1, y1, width, height)
         painter.drawPixmap(rect, self.pixmap, rect)
-        
+
     def setup_cr(self):
         cr = QPainter(self.pixmap)
         pen = QPen()
@@ -163,12 +163,12 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
         #    width, height, clr_wd, rem_px, ival)
 
         dist = self.rgbmap.get_dist()
-        
+
         j = ival; off = 0
         range_pts = []
         for i in range(256):
-            
-            wd = clr_wd    
+
+            wd = clr_wd
             if rem_px > 0:
                 j -= 1
                 if j == 0:
@@ -188,8 +188,8 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
                 # get inverse of distribution function and calculate value
                 # at this position
                 rng_pct = dist.get_dist_pct(cb_pct)
-                val = int(self.loval + (rng_pct * (self.hival - self.loval)))
-                text = "%d" % (val)
+                val = float(self.loval + (rng_pct * (self.hival - self.loval)))
+                text = "%.4g" % (val)
                 rect = cr.boundingRect(0, 0, 1000, 1000, 0, text)
                 x1, y1, x2, y2 = rect.getCoords()
                 _wd = x2 - x1
@@ -199,7 +199,7 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
                 rx = x
                 ry = _ht - 2
                 range_pts.append((rx, ry, text))
-                
+
             off += wd
 
         # draw range
@@ -231,7 +231,7 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
         ## rect = self.geometry()
         ## self.repaint(rect)
         self.update()
-        
+
     def shift_colormap(self, pct):
         self.rgbmap.set_sarr(self._sarr, callback=False)
         self.rgbmap.shift(pct)
@@ -239,7 +239,7 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
 
     def rgbmap_cb(self, rgbmap):
         self.redraw()
-    
+
     def mousePressEvent(self, event):
         buttons = event.buttons()
         x, y = event.x(), event.y()
@@ -248,7 +248,7 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
             self._start_x = x
             sarr = self.rgbmap.get_sarr()
             self._sarr = sarr.copy()
-                
+
     def mouseReleaseEvent(self, event):
         # note: for mouseRelease this needs to be button(), not buttons()!
         buttons = event.button()
@@ -278,7 +278,7 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
 
         width, height = self.get_size()
         pct = float(x) / float(width)
-        value = int(self.loval + pct * (self.hival - self.loval))
+        value = float(self.loval + pct * (self.hival - self.loval))
         self.make_callback('motion', value, event)
 
     def wheelEvent(self, event):
@@ -288,7 +288,7 @@ class ColorBar(Callback.Callbacks, QtGui.QWidget):
             direction = 'up'
         elif delta < 0:
             direction = 'down'
-            
+
         self.make_callback('scroll', event)
 
 #END
